# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Takker Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/me": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "description": "Returns user profile",
                "operationId": "37f901c9e1d2391b1cc4e4fa43cb91e8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/update-profile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update a current user profile and Settings",
                "operationId": "update",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "String",
                                            "example": "Profile updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "integer",
                                            "example": "The tap interval field must be a number."
                                        },
                                        "errors": {
                                            "properties": {
                                                "tap_interval": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The tap interval field must be a number"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/amz-sync": {
            "post": {
                "tags": [
                    "Amazon Sync"
                ],
                "summary": "User sync with amazon account",
                "operationId": "d4b0fc222804095768dc3a951ccff99e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "body",
                                "required": [
                                    "link"
                                ],
                                "properties": {
                                    "link": {
                                        "type": "string",
                                        "example": "Magical link from Amazon account"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Amazon Sync successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "title": "url",
                                            "type": "integer",
                                            "example": "The link field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "link": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The link field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AmzTokens": {
                "title": "AmzTokens",
                "description": "A amazon token taken from the Bot.",
                "properties": {
                    "amz_token": {
                        "title": "Amazon Token",
                        "description": "",
                        "type": "string",
                        "example": ""
                    },
                    "amz_refresh_token": {
                        "title": "Amazon Refresh Token",
                        "description": "",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AmzTokens"
                }
            },
...
            "Profile": {
                "title": "Profile",
                "description": "Profile data has created when the user make a registration. This property have all properties inside the Settings and Preferences",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 2
                    },
                    "amz_token": {
                        "title": "Amazon Token",
                        "description": "Amazon Token retrieved before login with Amazon Account",
                        "type": "string",
                        "example": ""
                    },
                    "tap_interval": {
                        "title": "Tap interval",
                        "description": "In Seconds",
                        "example": "6.2"
                    },
                    "auto_stop_after_crash": {
                        "title": "Auto stop after crash",
                        "description": "Auto stop after crash",
                        "type": "boolean",
                        "example": "false"
                    },
                    "auto_resume_search": {
                        "title": "Auto resume search",
                        "description": "Auto resume search",
                        "type": "boolean",
                        "example": "false"
                    },
                    "offer_lead_time": {
                        "title": "Offer Lead Time",
                        "description": "In Minutes",
                        "example": "15"
                    },
                    "minimum_base_paytype": {
                        "title": "Mininum Base Pay Type",
                        "description": "Mininum Base Pay Type",
                        "type": "string",
                        "example": "Per Offer"
                    },
                    "minimum_base_payvalue": {
                        "title": "Minimum Base Pay value",
                        "description": "In dollars",
                        "example": "15"
                    },
                    "offer_duration_start": {
                        "title": "Start Offer Duration",
                        "description": "",
                        "type": "string",
                        "example": "1:00"
                    },
                    "offer_duration_end": {
                        "title": "End Offer Duration",
                        "description": "",
                        "type": "string",
                        "example": "2:45"
                    },
                    "working": {
                        "title": "Bot Working",
                        "description": "Enable bot working",
                        "type": "boolean",
                        "example": "false"
                    },
                    "timezone": {
                        "title": "TimeZone in Profile",
                        "description": "",
                        "type": "string",
                        "example": "America/Brasilia"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Profile"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Username",
                        "description": "Username",
                        "type": "string",
                        "format": "int64",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "User Email",
                        "description": "Email address",
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "email_verified_at": {
                        "title": "Email verified",
                        "type": "string",
                        "format": "date-time",
                        "example": "null"
                    },
                    "created_at": {
                        "title": "CreatedAt",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "UpdatedAt",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    },
                    "amz_tokens": {
                        "title": "AmazonTokens",
                        "description": "",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AmzTokens"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "UpdateProfileRequest": {
                "title": "UpdateProfileRequest",
                "description": "Update Profile request body data",
                "properties": {
                    "amz_token": {
                        "title": "Amazon Token",
                        "type": "string",
                        "example": ""
                    },
                    "tap_interval": {
                        "title": "Tap interval",
                        "description": "In Seconds",
                        "type": "number",
                        "format": "float",
                        "example": "6.2"
                    },
                    "auto_stop_after_crash": {
                        "title": "Auto stop after crash",
                        "description": "Auto stop after crash",
                        "type": "boolean",
                        "example": "false"
                    },
                    "auto_resume_search": {
                        "title": "Auto resume search",
                        "description": "Auto resume search",
                        "type": "boolean",
                        "example": "false"
                    },
                    "offer_lead_time": {
                        "title": "Offer Lead Time",
                        "description": "In Minutes",
                        "type": "number",
                        "example": "15"
                    },
                    "minimum_base_paytype": {
                        "title": "Mininum Base Pay Type",
                        "description": "Mininum Base Pay Type",
                        "type": "string",
                        "example": "Per Offer"
                    },
                    "minimum_base_payvalue": {
                        "title": "Minimum Base Pay value",
                        "description": "In dollars",
                        "type": "number",
                        "example": "15"
                    },
                    "offer_duration_start": {
                        "title": "Start Offer Duration",
                        "description": "",
                        "type": "string",
                        "example": "1:00"
                    },
                    "offer_duration_end": {
                        "title": "End Offer Duration",
                        "description": "",
                        "type": "string",
                        "example": "2:45"
                    },
                    "working": {
                        "title": "Bot Working",
                        "description": "Enable bot working",
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateProfileRequest"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    }
}